backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
mysql-connector-python
python-dotenv